working richards reactive transport

+++ pfsimulator/parflow_lib/pf_alquimia.h
@@ -208,7 +208,7 @@ void PrintChemistryData(ChemPrintFlags *print_flags, AlquimiaSizes *chem_sizes, 
 
 void SelectReactTransTimeStep(double max_velocity, double CFL, 
                 double PF_dt, double *advect_react_dt, 
-                double *num_rt_iterations);
+                int *num_rt_iterations);
 
 
 #endif
